Refactor(Toast): Abstract Default-Style
